+++ client/src/js/database.js
@@ -1,21 +1,40 @@
 import { openDB } from 'idb';
 
+const JATE = 'jate'
+
 const initdb = async () =>
-  openDB('jate', 1, {
+  openDB(JATE, 1, {
     upgrade(db) {
-      if (db.objectStoreNames.contains('jate')) {
+      if (db.objectStoreNames.contains(JATE)) {
         console.log('jate database already exists');
         return;
       }
-      db.createObjectStore('jate', { keyPath: 'id', autoIncrement: true });
+      db.createObjectStore(JATE, { keyPath: 'id', autoIncrement: true });
       console.log('jate database created');
     },
   });
 
 // TODO: Add logic to a method that accepts some content and adds it to the database
-export const putDb = async (content) => console.error('putDb not implemented');
+export const putDb = async (content) => {
+  console.log('PUT to the database');
+  const jateDb = await openDB(JATE, 1);
+  const tx = jateDb.transaction(JATE, 'readwrite');
+  const store = tx.objectStore(JATE);
+  const request = store.put({ value: content });
+  const result = await request;
+  console.log('Data saved to the database', result);
+};
 
 // TODO: Add logic for a method that gets all the content from the database
-export const getDb = async () => console.error('getDb not implemented');
+export const getDb = async () => {
+  console.log('GET all from the database');
+  const jateDb = await openDB(JATE, 1);
+  const tx = jateDb.transaction(JATE, 'readonly');
+  const store = tx.objectStore(JATE);
+  const request = store.getAll();
+  const result = await request;
+  console.log('result.value', result);
+  return result;
+};
 
 initdb();