+++ client/webpack.config.js
@@ -56,9 +56,9 @@ module.exports = (env, argv) => {
       new WebpackPwaManifest({
         fingerprints: false,
         inject: true,
-        name: 'Contact Cards',
-        short_name: 'Contact',
-        description: 'Never forget your contacts!',
+        name: 'Just Another Text Editor',
+        short_name: 'JATE',
+        description: 'A PWA Text Editor',
         background_color: '#225ca3',
         theme_color: '#225ca3',
         start_url: './',