+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 1. Setup Folder Structure and loaded Starter Code
 2. Added scripts to container package.json & added htmlWebPackPlugi to webpack.config.js
-3. Updated package.json, index.js & webpack.config.json  
+3. Updated package.json, index.js & webpack.config.json 
+4. Added code to get all data from database and to adds data to database